+++ internal/resources/istio_resources_finder.go
@@ -3,7 +3,7 @@ package resources
 import (
 	"context"
 	"fmt"
-	moduleLabels "github.com/kyma-project/istio/operator/pkg/labels"
+	"github.com/kyma-project/istio/operator/pkg/labels"
 	"os"
 	"regexp"
 
@@ -90,8 +90,7 @@ func (i *IstioResourcesFinder) FindUserCreatedIstioResources() ([]Resource, erro
 			return nil, err
 		}
 		for _, item := range u.Items {
-			val, exists := item.GetLabels()[moduleLabels.ModuleLabelKey]
-			if exists && val == moduleLabels.ModuleLabelValue {
+			if labels.HasModuleLabels(item) {
 				continue
 			}
 			res := Resource{
